Fix an issue where pre-existing ZEPHYR_BASE prevents project creation
If a project is deleted without also deleting content on storage,
ZEPHYR_BASE still has its link which prevents the project from
created properly. So make sure it is deleted first.

This also fixes an issue where multiple Zephyr Application projects
are created. The validateLinkLocation() returns WARNING if the linked
resource already appears in the workspace (aka in another project).
So WARNING is not an issue.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>