Refactor build configuration and toolchain classes
This separates the build configurations into one for cmake,
and the other for actually building the application.
This move is to logically separate the two phases of
Zephyr's build system. Although the toolchain variant is
specified before CMake, the actual compiler binaries to be
used are not known until after CMake is invoked. This is
an intermediate step towards implementing classes for
Zephyr's toolchains.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>