Customize indexer scanner for Zephyr build system
By default, GCCToolChain uses the host GCC/Clang compilers to index
source files. This patch pulls related functions from GCCToolChain
into ZephyrApplicationToolChain, and are modified to work with
Zephyr's build system and the toolchain executables discovered by
CMake.

In addition, the files involved in building the applications are
scanned with the command lines compiling those individual files.
This allows the most accurate include paths and macros to be used
for indexing those files. These command lines and resulting
scanner info objects are cached as generic scanners which are
being used for scanning other files in the Zephyr base tree.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>