Split build configuration into Makefiles and Ninja
With one single build configuration comes one single object.
This causes issues when there are multiple projects with mixed
usage of CMake generators. The internal states of that build
configuration and toolchain are dictated by the latest build.
So it is possible the internal states point to Ninja but
a Makefiles-based project is going to build next, which results
in the wrong program to be called. So split the build
configuration and toolchain into Makefiles and Ninja to avoid
this issue.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>