Delay indexer until project setup is completed
If auto indexing is turned on, the indexer will start indexing
files immediately, even before all the source entries are
setup correctly. This causes the indexer to index everything
under the project without regard to the exclusion in source
entries. So delay the indexer until after all the source
entries have been setup.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>