Fix build not configured... again
There are times when CBuildConfigurationManager cannot get
the Zephyr build configurations at start-up (possibly due to
start-up initialization latencies of plugins), and the project
build configuration is forever blacklisted. This affects all
the build, run and debug operations relying on the manager
returning a working CBuildConfiguration. There is currently
no methods to remove a blacklisted configuration. So here
we are to workaround this by forcing access to the blacklist
and manipulate it ourselves.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>