Add primitive debugging support
This is the first attempt to add debugging support for Zephyr
Applications. New debug configurations are added so flashing
the hardware and starting debugserver can be done before
launching the debugging session. However, setting up GDB is
not entirely automatic as developers need to enter the TCP/IP
port number of the debugserver similar to Eclipse CDT's own
"GDB Hardware Debugging" configuration.

Also simplifies the dependencies in the target file so maven
has one few remote site to contact.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>