Workaround process killing with Ninja
With Java 8, there is no way to figure out all the descendants
started by Ninja. Although the previous hack finds the actual
commands started by Ninja and runs them, this does not totally
work when Ninja is calling West, as West is also spawning
processes on its own. Sending a TERM signal to West does not
always terminate all the descendants. So we need to find all
the lowest level descendants of the process tree started by
Ninja and terminate them manually. This eventually will cause
Ninja to stop itself as all its descendants have been terminated.

On Linux and macOS, pgrep is used to find all the leaf nodes.
On Windows, taskkill makes it "easier" to kill by window titles.
So we can spawn a new process with our custom title. Once we
are done with it, use taskkill to find those windows and kill
the processes associated with that.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>