Add Java 11 specific process launching code
Since Java 9, java.lang.Process finally has built-in support for
getting children of a process, thus avoiding any hacks to destroy
those children. This adds a fragment to utilize this capability,
and will be used if Eclipse runs with Java 11 or later.

With this fragment, building the plugin now requires Java SE 11.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>