Add a generic toolchain class
This adds a generic toolchain class which can be used
in the future for non-GCC compatible toolchain.

But at this moment, it is being used to provide a default
C build configuration if the CBuildConfigurationManager
is asking for a default one. If this happens, the project
configuration is probably broken so we need a minimally
working build configuration. This allows the project to be
built without erroring out all the time.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>