Don't fix build configuration
It is observed that Eclipse likes to call ZephyrApplicationBuilder
on a project without a build configuration. When this happens,
we used to create a build configuration manually and continued
the build. However, the CDT CBuilder simply returns null in this
situation, so we are following the CDT way. Eclipse will eventually
call our builder with a properly configured project, as observed.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>