Update Debugger tab in debug launch configuration
When changing the JTAG device without saving first, the host
and port of target device would revert back to default of
the associated Java class, even though the tab is created with
different defaults for runners. In order to fix this, we have
to publish our own JTAG devices with our own defaults.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>