Always use Zephyr build configuration when building/launch
The projects' active build configurations do not always point to
the Zephyr build configurations even though it is specified when
the projects are first created. It happens to many times that
CDT would set ErrorBuildConfiguration as the active build
configurations. So for Zephyr application builder and launchbar,
don't rely on Eclipse/CDT to give us the correct configuration,
but to find one our own.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>