Delay indexer action until the first build is done
Given that compiler is not known until CMake is run
after the very first build after project creation
and that indexer would not work until the compiler
is identified, to avoid indexer from running against
the whole Zephyr source tree, all files are excluded
when project is first created. When doing a full
build, the exclusion list is updated to a standard
one, so that indexer will be able to work correctly.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>