build: fix getting cmake_c_compiler from cache
CMakeCache.txt is showing CMAKE_C_COMPILER as a STRING instead of
FILEPATH the first time that we build from plugin. When plugin retrieves
cache information, it is failing to recover CMAKE_C_COMPILER, and
triggering an exception during build. Searching for STRING if FILEPATH
is not found solves the issue.

Signed-off-by: Marta Navarro <marta.navarro@intel.com>